move non production method back to dev script
